transformer api refactor
