+++ python/llm/example/transformers/int4_pipeline.py
@@ -18,41 +18,24 @@ import time
 import argparse
 
 
-def convert_and_load(repo_id_or_model_path, model_family, n_threads):
+def convert(repo_id_or_model_path, model_family, tmp_path):
+    from bigdl.llm import llm_convert
+    original_llm_path = repo_id_or_model_path
+    bigdl_llm_path = llm_convert(
+        model=original_llm_path,
+        outfile='./',
+        outtype='int4',
+        tmp_path=tmp_path,
+        model_family=model_family)
 
-    from bigdl.llm.ggml.transformers import AutoModelForCausalLM
+    return bigdl_llm_path
 
-    # here you may input the HuggingFace repo id directly as the value of `pretrained_model_name_or_path`.
-    # This will allow the pre-trained model to be downloaded directly from the HuggingFace repository.
-    # The downloaded model will then be converted to binary format with int4 dtype weights,
-    # and saved into the cache_dir folder.
-    #
-    # if you already have the pre-trained model downloaded, you can provide the path to
-    # the downloaded folder as the value of `pretrained_model_name_or_path``
-    llm = AutoModelForCausalLM.from_pretrained(
-        pretrained_model_name_or_path=repo_id_or_model_path,
+def load(model_path, model_family, n_threads):
+    from bigdl.llm.transformers import BigdlForCausalLM
+    llm = BigdlForCausalLM.from_pretrained(
+        pretrained_model_name_or_path=model_path,
         model_family=model_family,
-        dtype='int4',
-        cache_dir='./',
         n_threads=n_threads)
-
-    # if you want to explicitly convert the pre-trained model, you can use the `llm_convert` API 
-    # to convert the downloaded Huggungface checkpoint first,
-    # and then load the binary checkpoint directly.
-    #
-    # from bigdl.llm import llm_convert
-    #
-    # model_path = repo_id_or_model_path
-    # output_ckpt_path = llm_convert(
-    #     model=model_path,
-    #     outfile='./',
-    #     outtype='int4',
-    #     model_family=model_family)
-    #
-    # llm = AutoModelForCausalLM.from_pretrained(
-    #     pretrained_model_name_or_path=output_ckpt_path,
-    #     model_family=model_family,
-    #     n_threads=n_threads)
 
     return llm
 
@@ -62,27 +45,27 @@ def inference(llm, repo_id_or_model_path, model_family, prompt):
         # Option 1: Use HuggingFace transformers tokenizer
         print('-'*20, ' HuggingFace transformers tokenizer ', '-'*20)
         
-        print('Please note that the loading of HuggingFace transformers tokenizer may take some time.\n')
+        # print('Please note that the loading of HuggingFace transformers tokenizer may take some time.\n')
         # here is only a workaround for default example model 'decapoda-research/llama-7b-hf' in LLaMA family,
         # due to its out-of-date 'tokenizer_class' defined in its tokenizer_config.json.
         #
         # for most cases, you could use `AutoTokenizer`.
-        if model_family == 'llama':
-            from transformers import LlamaTokenizer
-            tokenizer = LlamaTokenizer.from_pretrained(repo_id_or_model_path)
-        else:
-            from transformers import AutoTokenizer
-            tokenizer = AutoTokenizer.from_pretrained(repo_id_or_model_path)
+        # if model_family == 'llama':
+        #     from transformers import LlamaTokenizer
+        #     tokenizer = LlamaTokenizer.from_pretrained(repo_id_or_model_path)
+        # else:
+        #     from transformers import AutoTokenizer
+        #     tokenizer = AutoTokenizer.from_pretrained(repo_id_or_model_path)
 
-        st = time.time()
+        # st = time.time()
 
-        # please note that the prompt here can either be a string or a list of string
-        tokens_id = tokenizer(prompt).input_ids
-        output_tokens_id = llm.generate(tokens_id, max_new_tokens=32)
-        output = tokenizer.batch_decode(output_tokens_id)
+        # # please note that the prompt here can either be a string or a list of string
+        # tokens_id = tokenizer(prompt).input_ids
+        # output_tokens_id = llm.generate(tokens_id, max_new_tokens=32)
+        # output = tokenizer.batch_decode(output_tokens_id)
 
-        print(f'Inference time: {time.time()-st} s')
-        print(f'Output:\n{output}')
+        # print(f'Inference time: {time.time()-st} s')
+        # print(f'Output:\n{output}')
 
         # Option 2: Use bigdl-llm based tokenizer
         print('-'*20, ' bigdl-llm based tokenizer ', '-'*20)
@@ -121,6 +104,8 @@ def main():
                              ', or the path to the huggingface checkpoint folder')
     parser.add_argument('--prompt', type=str, default='Q: What is CPU? A:',
                         help='Prompt to infer')
+    parser.add_argument('--tmp-path', type=str, default='/tmp',
+                        help='path to store intermediate model during the conversion process')
     args = parser.parse_args()
 
     repo_id_or_model_path = args.repo_id_or_model_path
@@ -132,12 +117,18 @@ def main():
         elif args.model_family == 'bloom':
             repo_id_or_model_path = 'bigscience/bloomz-7b1'
 
-    # Step 1: convert and load int4 model
-    llm = convert_and_load(repo_id_or_model_path=repo_id_or_model_path,
-                           model_family=args.model_family,
-                           n_threads=args.thread_num)
+    # Step 1: convert original model to BigDL llm model
+    bigdl_llm_path = convert(repo_id_or_model_path=repo_id_or_model_path,
+                               model_family=args.model_family,
+                               tmp_path=args.tmp_path)
+    
+    
+    # Step 2: load int4 model
+    llm = load(model_path=bigdl_llm_path,
+               model_family=args.model_family,
+               n_threads=args.thread_num)
 
-    # Step 2: conduct inference
+    # Step 3: inference
     inference(llm=llm,
               repo_id_or_model_path=repo_id_or_model_path,
               model_family=args.model_family,
