+++ python/llm/src/bigdl/llm/transformers/__init__.py
@@ -16,3 +16,4 @@
 
 from .convert import ggml_convert_int4
 from .model import AutoModelForCausalLM, AutoModel
+from .modelling_bigdl import BigdlForCausalLM
